Correciones en el modulo Productos

+++ Backend/models/movimiento.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Numeric
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from database import Base
@@ -10,7 +10,7 @@ class Movimiento(Base):
     id_producto = Column(Integer, ForeignKey("producto.id_producto"), nullable=False)
     tipo = Column(String, nullable=False)  # 'Entrada' o 'Salida'
     origen = Column(String, nullable=False)  # 'Scanner' o 'Manual'
-    cantidad = Column(Integer, nullable=False)
+    cantidad = Column(Numeric(10, 2), nullable=False)
     fecha_hora = Column(DateTime(timezone=True), server_default=func.now())
 
     producto = relationship("Producto")
